requirements.txt
numpy
pillow
matplotlib
reedsolo